Rename module in ext_classes test
Fixes #229
